Simplify image handling hierarchy
AbstractApplicationImage had nothing abstract in it, actually it
contains the whole implementation. Merge it into ApplicationImage and
remove the single method from ApplicationImage adjustLegacyPath that
seems to no longer be needed.
